+++ python/src/ux_channel/static/__init__.py
@@ -1,0 +1,4 @@
+# MANUAL_PUBLIC_API
+"""Static assets package."""
+PACKAGE = "static"
+__all__ = ["PACKAGE"]